+++ src/meshAMR/meshAMRinternalSurface_NastranSurface.h
@@ -35,14 +35,14 @@ public:
   typedef void (*fPrintTitle)(FILE*);
   fPrintTitle PrintTitle;
 
-  typedef void (*fPrintDataStateVector)(FILE* fout,int nElement,int *SurfaceElementsInterpolationList,int SurfaceElementsInterpolationListLength,cInternalNastranSurfaceData *Sutface,int spec,CMPI_channel* pipe,int ThisThread,int nTotalThreads);
+  typedef void (*fPrintDataStateVector)(FILE* fout,long int nElement,long int *SurfaceElementsInterpolationList,long int SurfaceElementsInterpolationListLength,cInternalNastranSurfaceData *Sutface,int spec,CMPI_channel* pipe,int ThisThread,int nTotalThreads);
   fPrintDataStateVector PrintDataStateVector;
 
   typedef double (*fLocalResolution)(double *);
   fLocalResolution localResolution;
 
   #if _AMR_DEBUGGER_MODE_ == _AMR_DEBUGGER_MODE_ON_
-  int Temp_ID;
+  long int Temp_ID;
   #endif
 
   void cleanDataBuffer() {
@@ -66,7 +66,7 @@ public:
 
 
   void GetSurfaceElementIndex() {}
-  int GetTotalSurfaceElementsNumber() {return 1;}
+  long int GetTotalSurfaceElementsNumber() {return 1;}
   double GetSurfaceElementArea(int nPolarElement) { return 1.0;}
 
   void PrintSurfaceData(const char *fname,int nDataSet, bool PrintStateVectorFlag=true) {
